CR-760 restructure CaseServiceImpl unit tests so one test class per method tested, since the test class was getting too big and cumbersome
