+++ src/test/java/uk/gov/ons/ctp/integration/contactcentresvc/service/impl/CaseServiceImplTestBase.java
@@ -1,0 +1,53 @@
+package uk.gov.ons.ctp.integration.contactcentresvc.service.impl;
+
+import static org.junit.Assert.assertTrue;
+
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import ma.glasnost.orika.MapperFacade;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Spy;
+import uk.gov.ons.ctp.common.event.EventPublisher;
+import uk.gov.ons.ctp.common.time.DateTimeUtil;
+import uk.gov.ons.ctp.integration.caseapiclient.caseservice.CaseServiceClientServiceImpl;
+import uk.gov.ons.ctp.integration.common.product.ProductReference;
+import uk.gov.ons.ctp.integration.contactcentresvc.CCSvcBeanMapper;
+import uk.gov.ons.ctp.integration.contactcentresvc.config.AppConfig;
+import uk.gov.ons.ctp.integration.contactcentresvc.repository.CaseDataRepository;
+import uk.gov.ons.ctp.integration.contactcentresvc.service.AddressService;
+import uk.gov.ons.ctp.integration.contactcentresvc.service.CaseService;
+import uk.gov.ons.ctp.integration.eqlaunch.service.EqLaunchService;
+import uk.gov.ons.ctp.integration.eqlaunch.service.impl.EqLaunchServiceImpl;
+
+public abstract class CaseServiceImplTestBase {
+  @Spy AppConfig appConfig = new AppConfig();
+
+  @Mock ProductReference productReference;
+
+  @Mock CaseServiceClientServiceImpl caseServiceClient;
+
+  @Mock EqLaunchService eqLaunchService = new EqLaunchServiceImpl();
+
+  @Mock EventPublisher eventPublisher;
+
+  @Spy MapperFacade mapperFacade = new CCSvcBeanMapper();
+
+  @Mock CaseDataRepository dataRepo;
+
+  @Mock AddressService addressSvc;
+
+  @InjectMocks CaseService target = new CaseServiceImpl();
+
+  void verifyTimeInExpectedRange(long minAllowed, long maxAllowed, Date dateTime) {
+    long actualInMillis = dateTime.getTime();
+    assertTrue(actualInMillis + " not after " + minAllowed, actualInMillis >= minAllowed);
+    assertTrue(actualInMillis + " not before " + maxAllowed, actualInMillis <= maxAllowed);
+  }
+
+  long asMillis(String datetime) throws ParseException {
+    SimpleDateFormat dateParser = new SimpleDateFormat(DateTimeUtil.DATE_FORMAT_IN_JSON);
+    return dateParser.parse(datetime).getTime();
+  }
+}
