+++ src/test/resources/uk/gov/ons/ctp/integration/contactcentresvc/service/impl/CaseServiceImplGetCaseByUprnTest.CachedCase.json
@@ -1,0 +1,15 @@
+{
+  "id": "b7565b5e-1396-4965-91a2-918c0d3642ed",
+  "uprn": "100041045018",
+  "createdDateTime": "2020-05-11T11:55:23.195Z",
+  "formattedAddress": "39 Sandford Walk, Exeter, EX1 2ET",
+  "addressLine1": "39 Sandford Walk",
+  "addressLine2": "",
+  "addressLine3": "",
+  "townName": "Exeter",
+  "postcode": "EX1 2ET",
+  "addressType": "HH",
+  "caseType": "HH",
+  "estabType": "Household",
+  "region": "E"
+}
