+++ src/test/resources/uk/gov/ons/ctp/integration/contactcentresvc/service/impl/CaseServiceImplLaunchTest.CaseContainerDTO.json
@@ -1,0 +1,96 @@
+[
+{
+  "id": "b7565b5e-1396-4965-91a2-918c0d3642ed",
+  "estabType": "Household",
+  "uprn": "334999999999",
+  "caseRef": "1000000000000001",
+  "caseType": "HH",
+  "createdDateTime": "2019-05-14T16:11:41.343+01:00",
+  "lastUpdated": "2019-05-14T16:15:32.221+01:00",
+  "addressLine1": "Napier House",
+  "addressLine2": "11 Park Street",
+  "addressLine3": "Parkhead",
+  "townName": "Glasgow",
+  "postcode": "G1 2AA",
+  "organisationName": "ON",
+  "addressLevel": "E",
+  "abpCode": "AACC",
+  "region": "E12345678",
+  "latitude": "41.40338",
+  "longitude": "2.17403",
+  "oa": "EE22",
+  "lsoa": "x1",
+  "msoa": "x2",
+  "lad": "H1",
+  "secureEstablishment": false,
+  "caseEvents": [
+    {
+      "id": "101",
+      "eventType": "CASE_CREATED",
+      "description": "Initial creation of case",
+      "createdDateTime": "2019-05-14T16:11:41.343+01:00"
+    },
+    {
+      "id": "102",
+      "eventType": "X11",
+      "description": "Should be filtered out",
+      "createdDateTime": "2019-05-15T17:02:12.343+04:00"
+    },
+    {
+      "id": "103",
+      "eventType": "CASE_UPDATED",
+      "description": "Create Household Visit",
+      "createdDateTime": "2019-05-16T12:12:12.343Z"
+    }
+  ],
+  "collectionExerciseId": "22684ede-7d5f-4f53-9069-2398055c61b2",
+  "surveyType": "x"
+},
+{
+  "id": "b7565b5e-2222-2222-2222-918c0d3642ed",
+  "estabType": "Household",
+  "estabUprn": "334111111111",
+  "uprn": "334999999999",
+  "caseRef": "103300000000001",
+  "caseType": "CE",
+  "createdDateTime": "2019-05-15T18:09:41.343+01:00",
+  "lastUpdated": "2019-05-14T16:15:32.221+01:00",
+  "addressLine1": "3 High Street",
+  "addressLine2": "Burnsby",
+  "addressLine3": "Bigtown",
+  "townName": "Glumby",
+  "postcode": "G1 6AA",
+  "organisationName": "OX",
+  "addressLevel": "E",
+  "abpCode": "XX",
+  "region": "E",
+  "latitude": "5.40338",
+  "longitude": "6.17403",
+  "oa": "EE2",
+  "lsoa": "xlsoa",
+  "msoa": "xmsoa",
+  "lad": "H2",
+  "secureEstablishment": true,
+  "caseEvents": [
+    {
+      "id": "101",
+      "eventType": "CASE_CREATED",
+      "description": "Initial creation of case",
+      "createdDateTime": "2019-05-14T16:11:41.343+01:00"
+    },
+    {
+      "id": "102",
+      "eventType": "X11",
+      "description": "Should be filtered out",
+      "createdDateTime": "2019-05-15T17:02:12.343+04:00"
+    },
+    {
+      "id": "103",
+      "eventType": "CASE_UPDATED",
+      "description": "Create Household Visit",
+      "createdDateTime": "2019-05-16T12:12:12.343Z"
+    }
+  ]
+}
+]
+
