+++ src/test/resources/uk/gov/ons/ctp/integration/contactcentresvc/service/impl/CaseServiceImplgetCaseByUprnTest.AddressIndexAddressCompositeDTO.json
